repackages message types
