+++ network/p2p/p2pconf/errors_test.go
@@ -5,11 +5,13 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+
+	"github.com/onflow/flow-go/network/p2p"
 )
 
 // TestErrInvalidLimitConfigRoundTrip ensures correct error formatting for ErrInvalidLimitConfig.
 func TestErrInvalidLimitConfigRoundTrip(t *testing.T) {
-	controlMsg := CtrlMsgGraft
+	controlMsg := p2p.CtrlMsgGraft
 	limit := uint64(500)
 
 	e := fmt.Errorf("invalid rate limit value %d must be greater than 0", limit)
